Added test for GetTemplateFile.

+++ TemplateHandlerTests/TemplateWriterTests.cs
@@ -1,7 +1,6 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Security.Cryptography;
+using System.Text;
 using TemplateHandler;
-using System.Resources;
-
 
 namespace TemplateHandlerTests
 {
@@ -10,11 +9,29 @@ namespace TemplateHandlerTests
     {
         private string testRelPath { get; set; }
 
+        private string getHasAsString(byte[] hash)
+        {
+            StringBuilder Sb = new StringBuilder();
+            foreach (byte b in hash)
+            {
+                Sb.Append(b.ToString("x2"));
+            }
+
+            return Sb.ToString();
+        }
+
+
+
+
+
         [TestCleanup()]
         public void Cleanup()
         {
             string masterDocumentsDir = Path.Combine(Directory.GetCurrentDirectory(), "MasterDocuments");
-            Directory.Delete(masterDocumentsDir, true);
+            if (Directory.Exists(masterDocumentsDir))
+            {
+                Directory.Delete(masterDocumentsDir, true);
+            }
         }
 
         [TestInitialize()]
@@ -35,7 +52,7 @@ namespace TemplateHandlerTests
 
 
         [TestMethod]
-        public void TestTemplateWriterInsertTemplate()
+        public void TestTemplateWriterInsertTemplateCorrect()
         {
             ArchiveFile file = new ArchiveFile(1, "38479384dhfhedfheiufh", testRelPath,
                                                 "WEIIHFRIWEHF136287136821634", "fmt/40", "Word file", 1, 200, null);
@@ -51,9 +68,33 @@ namespace TemplateHandlerTests
         }
 
         [TestMethod()]
-        public void GetTemplateFileTest()
+        public void GetTemplateFileTestValidID()
         {
-            Assert.Fail();
+            // Setup
+            int testID = 1;
+            SHA256 sha256 = SHA256.Create();
+
+            byte[] templateContent = TemplateWriter.GetTemplateFile(testID);
+            byte[] expectedContent = File.ReadAllBytes("Images/file_empty.tif");
+
+            // Compute the hash of th two files.
+            byte[] templateHash = sha256.ComputeHash(templateContent);
+            byte[] expectedHash = sha256.ComputeHash(expectedContent);
+
+            // Convert the hash to a string.
+            string templateHashString = getHasAsString(templateHash);
+            string expectedHashString = getHasAsString(expectedHash);
+
+            // Assert that the string hahshes are equal.
+            Assert.AreEqual(expectedHashString, templateHashString);
+
+        }
+
+        [TestMethod()]
+        public void GetTemplateFileTestNonValidID()
+        {
+            int testID = 500;
+            Assert.ThrowsException<ArgumentException>(() => TemplateWriter.GetTemplateFile(testID));
         }
     }
 }
